Use FieldOptionsList on SelectionField
